refactor : apply step1 feedbacks - component folder structure
