+++ .eslintrc.json
@@ -15,7 +15,8 @@
     "import/no-extraneous-dependencies": "off",
     "react/function-component-definition": "off",
     "jsx-a11y/click-events-have-key-events": "off",
-    "jsx-a11y/no-static-element-interactions": "off"
+    "jsx-a11y/no-static-element-interactions": "off",
+    "import/prefer-default-export": "off"
   },
   "parserOptions": {
     "project": "./tsconfig.json"
