+++ .github/workflows/chromatic.yml
@@ -1,0 +1,16 @@
+name: 'React Payments Storybook - yanggak12'
+
+# Event for the workflow
+on: push
+
+# List of jobs
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v1
+      - run: yarn
+      - uses: chromaui/action@v1
+        with:
+          projectToken: ${{ secrets.CHROMATIC_PROJECT_TOKEN }}
+          token: ${{ secrets.GITHUB_TOKEN }}
