+++ src/components/Buttons/index.ts
@@ -1,0 +1,3 @@
+import TextButton from './TextButton';
+
+export { TextButton };
