+++ src/components/Forms/index.ts
@@ -1,0 +1,4 @@
+import CardPasswordInput from './CardPasswordInput';
+import NewCardInput from './NewCardInput';
+
+export { CardPasswordInput, NewCardInput };
