+++ src/components/Icons/index.ts
@@ -1,4 +1,3 @@
-/* eslint-disable import/prefer-default-export */
 import LeftArrow from './LeftArrow';
 
 export { LeftArrow };
