+++ src/components/Modals/index.ts
@@ -1,0 +1,4 @@
+import BottomSheet from './BottomSheet';
+import BankBottomSheet from './BottomSheet/BankBottomSheet';
+
+export { BottomSheet, BankBottomSheet };
