+++ src/utils/functions.ts
@@ -1,4 +1,3 @@
-/* eslint-disable import/prefer-default-export */
 export const JsonToArr = <T>(Json: string | null) =>
   JSON.parse(Json as string) as T;
